fork-to-stulto - fork to stulto; license proceeding code under GPL
I'm not normally in favor of huge commits like this, but as far as I'm concerned, this may as well be the initial commit of a new project, so I'm less concerned about it here.
Also, it took, like, weeks of battle with the same few small deallocation bugs due to my C n00bishness, so as far as tidiness goes, for the moment I'm over it.