config-cleanup - remove current references to headerbar flag
We don't want this living in the terminal config, because it can't be
set on a per-terminal basis. To address this coupling (while
unfortunately rolling back functionality), we pull out references to the
setting completely and will eventually pick up `notify` signals to
update the window title and tab state in a way that keeps that logic
untangled from the concerns of unrelated widgets.