config-cleanup - StultoTerminalConfig: parse `options` section in module
This should simply not be coupled up with the parsing of command-line
options - if a user wants to override settings, they can use a new
profile for that.

As a compromise and to cut down on redundant copying, we'll probably
enforce the use of stulto.ini as a default config and enable partially
overwriting its settings with additional profiles.