terminal-keybindings-cleanup - StultoTerminal: remove obsolete font resize functions
Seems a bit hypocritical in light of the fact that the previous commit
mentions retaining font resizability outside of the app window, but if
we have keybindings to handle that work, we don't really need to
re-expose this capability via functions at the moment.

We'll reimplement the functions if it ever matters, but hopefully it
doesn't.