StultoHeaderBar: remove explicit `has-subtitle` set
This is partly an aesthetic choice, but also partly utilitarian. Given
that we've removed the terminal titlebar, the window Chroms is not as
heavy, so this looks less out of place (and a little more attractive
imo, but that's pure opinion).

On the other hand, we probably didn't want this set in the first place,
as the headerbar needs to be a little heavy to accommodate the buttons
and make them easy to access.